Rename asm segment to section
