+++ src/splat/segtypes/wasm/__init__.py
@@ -1,2 +1,2 @@
-from . import asm as asm
+from . import section as section
 from . import header as header
